latest update

+++ src/pages/welcome/welcome.ts
@@ -1,5 +1,5 @@
 import { Component, AfterViewInit} from '@angular/core';
-import { NavController,Events } from 'ionic-angular';
+import { NavController,Events,ToastController, LoadingController } from 'ionic-angular';
 
 import { LoginPage } from '../login/login';
 import { SignupPage } from '../signup/signup';
@@ -25,27 +25,48 @@ export class WelcomePage {
     "password":""
   }
   confirmationResult: any
-  constructor(public events:Events,public navCtrl: NavController, public fire : FirebaseService) {
+  constructor(public loadCtrl:LoadingController,public tc:ToastController,public events:Events,public navCtrl: NavController, public fire : FirebaseService) {
     this.fbs=fire
     this.ev=events
    }
 
    ngAfterViewInit(){
-     this.recaptchaVerifier=this.fbs.recaptcha("login-button")
-     this.recaptchaVerifier1=this.fbs.recaptcha("signup-button")
    }
+
   login() {
+
     var navCtrl=this.navCtrl
     var vm=this
     //navCtrl.push(LoginPage,{'confirm':"what?"})
-    this.fbs.login(this.creds, this.recaptchaVerifier).then(function(res){
-      console.log("We have a response: ", res)
-      navCtrl.push(LoginPage,{"confirm":res})
-    }).catch(function(err){
-    //  vm.fbs.currentUser().delete()
-      console.log("Error loging in. Cause: ",err)
-      //navCtrl.push(LoginPage)
+    var load1=this.loadCtrl.create({
+      content:"Sending you your SMS code..."
     })
+
+    load1.present()
+    this.fbs.recaptcha("login-button").then(value => {
+      console.log("success reCaptcha")
+      vm.recaptchaVerifier=value
+      vm.fbs.login(vm.creds, vm.recaptchaVerifier).then(function(res){
+        console.log("We have a response: ", res)
+        load1.dismiss()
+        navCtrl.push(LoginPage,{"confirm":res})
+      }).catch(function(err){
+      //  vm.fbs.currentUser().delete()
+        load1.dismiss()
+        console.log("Error loging in. Cause: ",err)
+        var toast=vm.tc.create({
+          message: "Sorry, an error has occured. Make sure you are connected to the internet, and that you have entered a valid Ethiopian phone number.",
+          duration: 3000,
+          position: 'top'
+        })
+        toast.present()
+        //navCtrl.push(LoginPage)
+      })
+    }).catch(function(err){
+      load1.dismiss()
+      console.log("error recaptcha ",err)
+    })
+
   }
 
   signup() {
@@ -54,20 +75,58 @@ export class WelcomePage {
     var creds= this.creds
     var vm=this
 
+
     var recaptchaVerifier1 = this.recaptchaVerifier1*/
-    var vm=this
-    var confirmationResult;
-    var navCtrl=this.navCtrl
-    this.fbs.createUser(this.creds, this.recaptchaVerifier1).then(function(res){
-      console.log("This is the conf code",res)
-      confirmationResult = res
-      navCtrl.push(SignupPage,{"confirm":confirmationResult})
-    }).catch(function(err){
-      vm.fbs.currentUser().delete()
-      console.log("You have an error",err)
-      //toast saying An error with the network occured, click back and try again.
-      //navCtrl.push(SignupPage,{"confirm":confirmationResult})
+    var load1=this.loadCtrl.create({
+      content:"Sending you your SMS code..."
     })
+    load1.present()
+
+    var vm=this
+    this.fbs.recaptcha("signup-button").then(value => {
+      console.log("success reCaptcha",value)
+      this.recaptchaVerifier1=value
+      var confirmationResult;
+      var navCtrl=this.navCtrl
+      vm.fbs.createUser(vm.creds, vm.recaptchaVerifier1).then(function(res){
+        console.log("This is the conf code",res)
+        load1.dismiss()
+        confirmationResult = res
+        navCtrl.push(SignupPage,{"confirm":confirmationResult})
+      }).catch(function(err){
+        load1.dismiss()
+        if(vm.fbs.currentUser()){
+          vm.fbs.currentUser().delete()
+        }
+        if(err==="notEthiopian"){
+          var toast=vm.tc.create({
+            message:"Please make sure you entered your Ethiopian phone number. e.g. 931605471 or 0931605471",
+            duration: 3000,
+            position: 'top'
+          })
+          toast.present()
+        }else{
+          var toast=vm.tc.create({
+            message: "Sorry, an error has occured. Make sure you are connected to the internet.",
+            duration: 3000,
+            position: 'top'
+          })
+          toast.present()
+        }
+        //toast saying An error with the network occured, click back and try again.
+        //navCtrl.push(SignupPage,{"confirm":confirmationResult})
+      })
+    }).catch(function(err){
+      load1.dismiss()
+       var toast=vm.tc.create({
+        message: "Sorry, something went wrong in verifying your number. Please try again.",
+        duration: 3000,
+        position: 'top'
+      })
+      toast.present()
+      console.log("error recaptcha ",err)
+    })
+
 
 
   }
